good progress; factor apart
