+++ texnotes/baker-campbell-hausdorff.ltx
@@ -1766,6 +1766,6 @@ And we find we need
 \[\color{blue} 24G_4 = 6G_{3'} + 6G_{[13]} + 3G_{[2'1]} + 3G_{[1'2]} + G_{[[1'1]1]} + G_{[[12]1]}\]
 Now this is looking very nice. If I expand out the ``derivatives'' I see
 \[ 24G_4 = 6G_{[03]} + 6G_{[13]} + 3G_{[[02]1]} + 3G_{[[01]2]} + G_{[[[01]1]1]} + G_{[[12]1]}\]
-and apart from some confusion about $G_0$ vs. $G_1$ this is covering all compositions of $4$, and they're
-all left-nested.
+If I do $[0,\dash] \mapsto [\dash, 1]$ everywhere, then all compositions of $4$ are covered, and they
+all remain left-nested.
 \end{document}]
